Completed task 2
