+++ projects/queues/README.md
@@ -12,5 +12,5 @@
         - How does it fit into the overall system?
         - How does it work? (architectural overview, e.g. what each module's purpose is or how data flows through the firmware)
 -->
-# hello_world
+# queues
